Move extra highligher config to CodeBlockLanguageDetector
